+++ CHANGELOG.md
@@ -11,7 +11,9 @@ The [releases page](https://github.com/Jelmerro/Vieb/releases) also contains the
 but the list below contains much more technical details.
 The releases of Vieb aim to follow [semantic versioning](https://semver.org).
 
-## Unreleased
+## [4.4.0](https://github.com/Jelmerro/Vieb/compare/4.3.0...4.4.0) - 2021-04-14
+
+[code diff](https://github.com/Jelmerro/Vieb/compare/4.3.0...4.4.0) - [released builds](https://github.com/Jelmerro/Vieb/releases/tag/4.4.0)
 
 ### Added
 
@@ -39,8 +41,8 @@ The releases of Vieb aim to follow [semantic versioning](https://semver.org).
 
 ### Security
 
-- Electron 12.0.2 (unchanged)
-- Chromium 89.0.4389.90 (unchanged)
+- Electron 12.0.4 (was 12.0.2)
+- Chromium 89.0.4389.114 (was 89.0.4389.90)
 
 ## [4.3.0](https://github.com/Jelmerro/Vieb/compare/4.2.1...4.3.0) - 2021-04-05